refactor: replace LetterEditorLayout to Letter component, just name changed
