+++ app/step/3/components/Letter/Editor/index.js
@@ -1,0 +1,2 @@
+export * from "./ToolBar.jsx"
+export * from "./ToolBarSelector.jsx"
